+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - [#670](https://github.com/bumble-tech/appyx/pull/670) - Fixes ios lifecycle
 - [#673](https://github.com/bumble-tech/appyx/pull/673) – Fix canHandeBackPress typo
 - [#671](https://github.com/bumble-tech/appyx/issue/671) – Fix ui state saving issue
+- [#694](https://github.com/bumble-tech/appyx/pull/694) – Fix appyxComponent state saving issue
 
 ### Enhancement